Collectables: Coins, Speed and Jump
I created the system for all collectables

The speed boosts speed, the jump gives an extra jump  which resets when landing.

There are also 4 collectables that give coin value, each of different size and colour.

Alongside this I made a basic player movement and camera, Tomorrow is the Switch, Doors, Logic and Cutscenes.